+++ Program.cs
@@ -10,13 +10,13 @@ namespace Generics
         {
             Console.WriteLine("Enter three value\n");
             Console.Write("First value:");
-            int firstNo = Convert.ToInt32(Console.ReadLine());
+            double firstNo = Convert.ToDouble(Console.ReadLine());
 
             Console.Write("Second value:");
-            int SecondNo = Convert.ToInt32(Console.ReadLine());
+            double SecondNo = Convert.ToDouble(Console.ReadLine());
 
             Console.Write("Third value:");
-            int ThirdNo = Convert.ToInt32(Console.ReadLine());
+            double ThirdNo = Convert.ToDouble(Console.ReadLine());
 
             Max max = new Max();
             max.MaxInt(firstNo, SecondNo, ThirdNo);
